refactor(frontend): split web chat follow-up owners
